+++ gradle-plugin/src/main/java/com/anatawa12/modPatching/common/CurseDownloadingMod.kt
@@ -1,4 +1,4 @@
-package com.anatawa12.modPatching
+package com.anatawa12.modPatching.common
 
 interface CurseDownloadingMod : DownloadingMod {
     var id: String
