+++ gradle-plugin/src/main/java/com/anatawa12/modPatching/common/DownloadingMod.kt
@@ -1,0 +1,11 @@
+package com.anatawa12.modPatching.common
+
+/**
+ * this interface is not public for implementing
+ */
+interface DownloadingMod {
+    /**
+     * name of this mod.
+     */
+    var name: String
+}
