+++ gradle-plugin/src/main/java/com/anatawa12/modPatching/common/internal/Constants.kt
@@ -1,0 +1,8 @@
+package com.anatawa12.modPatching.common.internal
+
+import org.gradle.util.GUtil
+
+object CommonConstants {
+    val DOWNLOAD_MODS = GUtil.toLowerCamelCase("download mods")
+    val PREPARE_MODS = GUtil.toLowerCamelCase("prepare mods")
+}
