+++ gradle-plugin/src/main/java/com/anatawa12/modPatching/common/internal/Delegates.kt
@@ -1,4 +1,4 @@
-package com.anatawa12.modPatching.internal
+package com.anatawa12.modPatching.common.internal
 
 import kotlin.reflect.KProperty
